Customer password encoder and security filters 4-28-2025

+++ src/main/java/com/pilogix/authserver/repo/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends CrudRepository<UserEntity, Integer> {
+public interface UserRepo extends CrudRepository<UserEntity, Long> {
     UserEntity findByUsername(String username);
 }
